Refactoring git, hg and hg-git
