+++ src/setuptools_scm/git.py
@@ -1,28 +1,20 @@
+from datetime import datetime, date
+import os
+from os.path import isfile, join, samefile
+import warnings
+
 from .config import Configuration
 from .utils import do_ex, trace, require_command
 from .version import meta
-from datetime import datetime, date
-import os
-from os.path import isfile, join
-import warnings
-
-
-from os.path import samefile
-
+from .scm_workdir import Workdir
 
 DEFAULT_DESCRIBE = "git describe --dirty --tags --long --match *[0-9]*"
 
 
-class GitWorkdir:
+class GitWorkdir(Workdir):
     """experimental, may change at any time"""
 
     COMMAND = "git"
-
-    def __init__(self, path):
-        self.path = path
-
-    def do_ex(self, cmd):
-        return do_ex(cmd, cwd=self.path)
 
     @classmethod
     def from_potential_worktree(cls, wd):
@@ -92,109 +84,6 @@ class GitWorkdir:
         return self.do_ex(DEFAULT_DESCRIBE)
 
 
-class GitWorkdirHgClient(GitWorkdir):
-    COMMAND = "hg"
-
-    @classmethod
-    def from_potential_worktree(cls, wd):
-        require_command(cls.COMMAND)
-        root, _, ret = do_ex("hg root", wd)
-        if ret:
-            return
-        return cls(root)
-
-    def is_dirty(self):
-        out, _, _ = self.do_ex("hg id -T '{dirty}'")
-        return bool(out)
-
-    def get_branch(self):
-        branch, err, ret = self.do_ex("hg id -T {bookmarks}")
-        if ret:
-            trace("branch err", branch, err, ret)
-            return
-        return branch
-
-    def get_head_date(self):
-        date_part, err, ret = self.do_ex("hg log -r . -T {shortdate(date)}")
-        if ret:
-            trace("head date err", date_part, err, ret)
-            return
-        return datetime.strptime(date_part, r"%Y-%m-%d").date()
-
-    def is_shallow(self):
-        return False
-
-    def fetch_shallow(self):
-        pass
-
-    def get_hg_node(self):
-        node, _, ret = self.do_ex("hg log -r . -T {node}")
-        if not ret:
-            return node
-
-    def node(self):
-        hg_node = self.get_hg_node()
-        if hg_node is None:
-            return
-
-        git_node = None
-        with open(os.path.join(self.path, ".hg/git-mapfile"), "r") as file:
-            for line in file:
-                if hg_node in line:
-                    git_node, hg_node = line.split()
-                    break
-
-        if git_node is None:
-            # trying again after hg -> git
-            self.do_ex("hg gexport")
-            return self.node()
-
-        return git_node[:7]
-
-    def count_all_nodes(self):
-        revs, _, _ = self.do_ex("hg log -r 'ancestors(.)' -T '.'")
-        return len(revs) + 1
-
-    def default_describe(self):
-        """
-        Tentative to reproduce the output of
-
-        `git describe --dirty --tags --long --match *[0-9]*`
-
-        """
-        hg_tags, _, ret = self.do_ex(
-            "hg log -r reverse(ancestors(.)) -T {tags}{if(tags, ' ', '')}"
-        )
-        if ret:
-            return None, None, None
-        hg_tags = hg_tags.split()
-
-        git_tags = {}
-        with open(os.path.join(self.path, ".hg/git-tags"), "r") as file:
-            for line in file:
-                node, tag = line.split()
-                git_tags[tag] = node
-
-        # find the first hg tag which is also a git tag
-        # TODO: also check for match *[0-9]*
-        for tag in hg_tags:
-            if tag in git_tags:
-                break
-
-        out, _, ret = self.do_ex("hg log -r .::" + tag + " -T .")
-        if ret:
-            return None, None, None
-        distance = len(out) - 1
-
-        node = self.node()
-        desc = f"{tag}-{distance}-g{node}"
-
-        if self.is_dirty():
-            desc += "-dirty"
-
-        return desc, None, 0
-
-
 def warn_on_shallow(wd):
     """experimental, may change at any time"""
     if wd.is_shallow():
@@ -225,6 +114,8 @@ def parse(root, describe_command=None, pre_parse=warn_on_shallow, config=None):
 
     wd = GitWorkdir.from_potential_worktree(config.absolute_root)
     if wd is None:
+        from .hg_git import GitWorkdirHgClient
+
         wd = GitWorkdirHgClient.from_potential_worktree(config.absolute_root)
     if wd is None:
         return
